server/src/requirements.txt
absl-py==2.1.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
bs4==0.0.1
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cors==1.0.1
cycler==0.11.0
decorator==4.4.2
discord==1.0.1
discord.py==1.6.0
distlib==0.3.3
easyocr==1.7.1
filelock==3.3.1
firebase-admin==4.5.3
Flask==2.2.5
Flask-Cors==4.0.0
flatbuffers==23.5.26
fonttools==4.38.0
future==1.0.0
gevent==22.10.2
google-api-core>=1.34.1,<3.0dev
google-api-python-client>=2.52.0,<3.0dev
google-auth>=1.19.0,<3.0.0
google-auth-httplib2==0.1.0
google-auth-oauthlib>=0.5.2,<3.0.0
google-cloud-core>=1.4.1,<2.0dev
google-cloud-firestore==2.0.2
google-cloud-storage>=1.11.0,<3.0.0
google-cloud-translate==3.15.3
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.56.3
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.48.2
gunicorn==20.1.0
httplib2==0.19.0
idna==2.10
imageio==2.31.2
imageio-ffmpeg==0.4.9
importlib-metadata==4.8.1
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
lxml==4.6.3
MarkupSafe==2.1.5
matplotlib==3.5.3
moviepy==1.0.3
msgpack==1.0.2
multidict==5.1.0
networkx==2.6.3
ninja==1.11.1.1
numpy==1.21.6
oauthlib==3.2.0
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==20.9
pathlib==1.0.1
Pillow==9.5.0
platformdirs==2.4.0
proglog==0.1.10
proto-plus==1.23.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post5
pycparser==2.20
pyparsing==2.4.7
PyQt5==5.15.4
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==0.17.1
pytz==2021.1
PyWavelets==1.3.0
PyYAML==6.0.1
qt5-applications==5.15.2.2.2
requests==2.31.0
requests-file==2.0.0
requests-oauthlib==1.3.1
rsa==4.7.2
scikit-image==0.19.3
scipy==1.7.3
shapely==2.0.4
six==1.15.0
slackclient==2.9.4
soupsieve==2.2.1
SpeechRecognition==3.9.0
threadpoolctl==3.1.0
tifffile==2021.11.2
tldextract==4.0.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.66.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
virtualenv==20.8.1
Werkzeug==2.2.3
yarl==1.6.3
zipp==3.6.0
zope.event==5.0
zope.interface==6.3
